refreshed code for VS2017
* ran the project again after a LOOOONG time.
* updated visual studio files to work with 2017
* looks like I was working on delauny/voronoi code, and A* pathfinding.
and tons of 2D geometry

+++ game/agent.h
@@ -150,7 +150,7 @@ public:
 				V2f top = right.rotated((float)V_PI / 4);
 				V2f bottom = right.rotated((float)-V_PI / 4);
 				float rr = body.radius * body.radius;
-				float squareDiagonal = sqrt( rr + rr );
+				float squareDiagonal = (float)sqrt( rr + rr );
 				right = direction * squareDiagonal;
 				right += body.center;
 				top += body.center;
